requirements.txt
fastapi
uvicorn
line-bot-sdk
sqlalchemy
python-dotenv
jinja2
python-multipart
aiofiles
crcmod
requests
pytest
httpx
